capstone/requirements.txt
pandas
requests
apache-airflow
apache-airflow-providers-postgres
psycopg2-binary